postgresql-connector/tests/requirements-test.txt
# Copyright 2020 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#      http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

attrs==19.3.0
bcrypt==3.1.7
bleach==3.1.4
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9
Flask==1.1.2
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-core==1.3.0
google-cloud-trace==0.23.0
googleapis-common-protos==1.52.0
grpcio==1.30.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
more-itertools==8.4.0
opentelemetry-api==0.10b0
opentelemetry-exporter-cloud-trace==0.10b0
opentelemetry-ext-flask==0.10b0
opentelemetry-ext-sqlalchemy==0.10b0
opentelemetry-ext-wsgi==0.10b0
opentelemetry-instrumentation==0.10b0
opentelemetry-sdk==0.10b0
packaging==20.4
pluggy==0.13.1
protobuf==3.12.2
psycopg2==2.7.7
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.7
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.1
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.14.0
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.2.0
